+++ certbot-dns-route53/certbot_dns_route53/_internal/__init__.py
@@ -1,0 +1,1 @@
+Internal implementation of `~certbot_dns_route53.dns_route53` plugin.
