+++ certbot-dns-route53/certbot_dns_route53/authenticator.py
@@ -1,16 +1,16 @@
-"""Shim around `~certbot_dns_route53.dns_route53` for backwards compatibility."""
+"""Shim around `~certbot_dns_route53._internal.dns_route53` for backwards compatibility."""
 import warnings
 
 import zope.interface
 
 from certbot import interfaces
-from certbot_dns_route53 import dns_route53
+from certbot_dns_route53._internal import dns_route53
 
 
 @zope.interface.implementer(interfaces.IAuthenticator)
 @zope.interface.provider(interfaces.IPluginFactory)
 class Authenticator(dns_route53.Authenticator):
-    """Shim around `~certbot_dns_route53.dns_route53.Authenticator` for backwards compatibility."""
+    """Shim around `~certbot_dns_route53._internal.dns_route53.Authenticator` for backwards compatibility."""
 
     hidden = True
