+++ certbot-dns-route53/setup.py
@@ -51,7 +51,7 @@ setup(
     keywords=['certbot', 'route53', 'aws'],
     entry_points={
         'certbot.plugins': [
-            'dns-route53 = certbot_dns_route53.dns_route53:Authenticator',
+            'dns-route53 = certbot_dns_route53._internal.dns_route53:Authenticator',
             'certbot-route53:auth = certbot_dns_route53.authenticator:Authenticator'
         ],
     },
